# frontend/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Deploy Hub API
  version: 0.0.1
paths:
...
  /api/v1/findings/:
    get:
      operationId: v1_findings_list
      parameters:
      - in: query
        name: entity
        schema:
          type: string
          maxLength: 128
          minLength: 1
      - in: query
        name: severity
        schema:
          enum:
          - p1
          - p2
          - p3
          type: string
          minLength: 1
        description: |-
          * `p1` - P1
          * `p2` - P2
          * `p3` - P3
      - in: query
        name: state
        schema:
          enum:
          - open
          - acked
          - resolved
          - accepted
          type: string
          minLength: 1
        description: |-
          * `open` - Open
          * `acked` - Acked
          * `resolved` - Resolved
          * `accepted` - Accepted
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingListSnapshot'
          description: ''
  /api/v1/findings/{id}/:
    get:
      operationId: v1_findings_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - v1
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingDetailSnapshot'
          description: ''
...
components:
  schemas:
...
    Finding:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        source_engine:
          type: string
          maxLength: 64
        severity:
          $ref: '#/components/schemas/SeverityEnum'
        entity:
          type: string
          maxLength: 128
        title:
          type: string
          maxLength: 256
        body:
          type: string
        fix_action:
          type: string
          maxLength: 256
        state:
          $ref: '#/components/schemas/StateEnum'
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        fingerprint:
          type: string
          maxLength: 128
        accepted_reason:
          type: string
          maxLength: 256
      required:
      - entity
      - fingerprint
      - id
      - severity
      - source_engine
      - title
    FindingDetailSnapshot:
      type: object
      properties:
        seq:
          type: integer
        data:
          $ref: '#/components/schemas/Finding'
      required:
      - data
      - seq
    FindingListSnapshot:
      type: object
      properties:
        seq:
          type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/Finding'
      required:
      - data
      - seq
...
    SeverityEnum:
      enum:
      - p1
      - p2
      - p3
      type: string
      description: |-
        * `p1` - P1
        * `p2` - P2
        * `p3` - P3
...
    StateEnum:
      enum:
      - open
      - acked
      - resolved
      - accepted
      type: string
      description: |-
        * `open` - Open
        * `acked` - Acked
        * `resolved` - Resolved
        * `accepted` - Accepted
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid